+++ src/platform/web/PAGWasmBindings.cpp
@@ -19,16 +19,16 @@
 #include <emscripten/bind.h>
 #include <emscripten/val.h>
 #include "base/utils/TGFXCast.h"
-#include "core/FontMetrics.h"
-#include "core/ImageInfo.h"
-#include "core/PathTypes.h"
-#include "gpu/opengl/GLDefines.h"
 #include "pag/pag.h"
 #include "pag/types.h"
 #include "platform/web/GPUDrawable.h"
 #include "platform/web/WebSoftwareDecoderFactory.h"
 #include "rendering/editing/StillImage.h"
 #include "rendering/video/VideoDecoder.h"
+#include "tgfx/core/FontMetrics.h"
+#include "tgfx/core/ImageInfo.h"
+#include "tgfx/core/PathTypes.h"
+#include "tgfx/gpu/opengl/GLDefines.h"
 #include "tgfx/src/platform/web/NativeImage.h"
 
 using namespace emscripten;
@@ -246,7 +246,7 @@ EMSCRIPTEN_BINDINGS(pag) {
                         auto origin = flipY ? ImageOrigin::BottomLeft : ImageOrigin::TopLeft;
                         return PAGSurface::MakeFrom(glTexture, origin);
                       }))
-      .class_function("_FromFrameBuffer",
+      .class_function("_FromRenderTarget",
                       optional_override([](int frameBufferID, int width, int height, bool flipY) {
                         GLFrameBufferInfo glFrameBufferInfo = {};
                         glFrameBufferInfo.id = static_cast<unsigned>(frameBufferID);
@@ -259,7 +259,14 @@ EMSCRIPTEN_BINDINGS(pag) {
       .function("_height", &PAGSurface::height)
       .function("_updateSize", &PAGSurface::updateSize)
       .function("_clearAll", &PAGSurface::clearAll)
-      .function("_freeCache", &PAGSurface::freeCache);
+      .function("_freeCache", &PAGSurface::freeCache)
+      .function("_readPixels",
+                optional_override([](PAGSurface& pagSurface, int colorType, int alphaType,
+                                     uintptr_t dstPixels, size_t dstRowBytes) {
+                  return pagSurface.readPixels(static_cast<ColorType>(colorType),
+                                               static_cast<AlphaType>(alphaType),
+                                               reinterpret_cast<void*>(dstPixels), dstRowBytes);
+                }));
 
   class_<PAGImage>("_PAGImage")
       .smart_ptr<std::shared_ptr<PAGImage>>("_PAGImage")
@@ -269,6 +276,23 @@ EMSCRIPTEN_BINDINGS(pag) {
       .class_function("_FromNativeImage", optional_override([](val nativeImage) {
                         return std::static_pointer_cast<PAGImage>(
                             StillImage::MakeFrom(tgfx::NativeImage::MakeFrom(nativeImage)));
+                      }))
+      .class_function("_FromPixels",
+                      optional_override([](uintptr_t pixels, int width, int height, size_t rowBytes,
+                                           int colorType, int alphaType) {
+                        return PAGImage::FromPixels(reinterpret_cast<void*>(pixels), width, height,
+                                                    rowBytes, static_cast<ColorType>(colorType),
+                                                    static_cast<AlphaType>(alphaType));
+                      }))
+      .class_function("_FromTexture",
+                      optional_override([](int textureID, int width, int height, bool flipY) {
+                        GLTextureInfo glInfo = {};
+                        glInfo.target = GL_TEXTURE_2D;
+                        glInfo.id = static_cast<unsigned>(textureID);
+                        glInfo.format = GL_RGBA8;
+                        BackendTexture glTexture(glInfo, width, height);
+                        auto origin = flipY ? ImageOrigin::BottomLeft : ImageOrigin::TopLeft;
+                        return PAGImage::FromTexture(glTexture, origin);
                       }))
       .function("_width", &PAGImage::width)
       .function("_height", &PAGImage::height)
@@ -362,8 +386,12 @@ EMSCRIPTEN_BINDINGS(pag) {
       .property("applyStroke", &TextDocument::applyStroke)
       .property("baselineShift", &TextDocument::baselineShift)
       .property("boxText", &TextDocument::boxText)
-      .property("boxTextPos", &TextDocument::boxTextPos)
-      .property("boxTextSize", &TextDocument::boxTextSize)
+      .property("boxTextPos", optional_override([](const TextDocument& textDocument) {
+                  return ToTGFX(textDocument.boxTextPos);
+                }))
+      .property("boxTextSize", optional_override([](const TextDocument& textDocument) {
+                  return ToTGFX(textDocument.boxTextSize);
+                }))
       .property("firstBaseLine", &TextDocument::firstBaseLine)
       .property("fauxBold", &TextDocument::fauxBold)
       .property("fauxItalic", &TextDocument::fauxItalic)
@@ -437,12 +465,6 @@ EMSCRIPTEN_BINDINGS(pag) {
       .value("EvenOdd", tgfx::PathFillType::EvenOdd)
       .value("InverseWinding", tgfx::PathFillType::InverseWinding)
       .value("InverseEvenOdd", tgfx::PathFillType::InverseEvenOdd);
-
-  enum_<ColorType>("ColorType")
-      .value("Unknown", ColorType::Unknown)
-      .value("ALPHA_8", ColorType::ALPHA_8)
-      .value("RGBA_8888", ColorType::RGBA_8888)
-      .value("BGRA_8888", ColorType::BGRA_8888);
 
   enum_<tgfx::LineCap>("LineCap")
       .value("Butt", tgfx::LineCap::Butt)